kafka_to_clickhouse/requirements.txt
kafka-python==2.0.2
clickhouse-driver==0.2.6
python-dotenv==1.0.0
pydantic==1.10.7
